requirement.txt
streamlit
pandas
numpy
plotly
pillow
base64
